requirements.txt
Django==1.8
django-braces==1.11.0
django-cors-middleware==1.3.1
django-extensions==1.7.9
django-filter==1.0.2
django-oauth-toolkit==0.12.0
django-post-office==2.0.8
djangorestframework==3.6.2
jsonfield==2.0.1
model-mommy==1.3.2
oauthlib==0.6.1
requests==2.14.2
urllib3==1.7.1
django-celery==3.2.1
celery==3.1.25
kombu==3.0.37
redis==2.10.5
